+++ TimeToCookTests/TimeToCookTests.swift
@@ -10,18 +10,6 @@ import XCTest
 
 var sut: ProductInfoViewModel?
 
-class ProductFake: ProductProtocol {
-    var weight: Int? = 200
-    var code = "12345678"
-    var title = "Title"
-    var producer = "Producer"
-    var category = "Рис"
-    var cookingTime = 20
-    var intoBoilingWater: Bool?
-    var needStirring: Bool?
-    var waterRatio = 3.0
-}
-
 class ProductInfoViewModelTests: XCTestCase {
     
     override func setUpWithError() throws {
@@ -36,15 +24,14 @@ class ProductInfoViewModelTests: XCTestCase {
     
     func testThatButtonStartCookTappedMustBeFalseAfterUpdateProduct() {
         // arrange
-        let productInfoViewModel = ProductInfoViewModel()
         let product: ProductProtocol? = nil
         
         // act
-        productInfoViewModel.buttonStartCookTapped = true
-        productInfoViewModel.updateProduct(product: product)
+        sut?.buttonStartCookTapped = true
+        sut?.updateProduct(product: product)
         
         // assert
-        XCTAssertFalse(productInfoViewModel.buttonStartCookTapped)
+        XCTAssertFalse(sut?.buttonStartCookTapped ?? true)
     }
     
     func testThatPropertyProductImageReturnCorrectImageName() {
@@ -52,10 +39,10 @@ class ProductInfoViewModelTests: XCTestCase {
         let product = ProductFake()
         
         // act
-        let productInfoViewModel = ProductInfoViewModel(product: product)
-        
+        sut?.updateProduct(product: product)
+
         // assert
-        XCTAssertEqual(productInfoViewModel.productImage, "Рис.png")
+        XCTAssertEqual(sut?.productImage, "Рис.png")
     }
     
     func testThatPropertyWeightReturnCorrectProductWeight() {
@@ -63,22 +50,22 @@ class ProductInfoViewModelTests: XCTestCase {
         let product = ProductFake()
         
         // act
-        let productInfoViewModel = ProductInfoViewModel(product: product)
+        sut?.updateProduct(product: product)
         
         // assert
-        XCTAssertEqual(productInfoViewModel.weight, "200 г.")
+        XCTAssertEqual(sut?.weight, "200 г.")
     }
     
     func testThatPropertyWeightReturnNotFoundIfProductIsNil() {
         // arrange
         let product = ProductFake()
-        let productInfoViewModel = ProductInfoViewModel(product: product)
+        sut?.updateProduct(product: product)
         
         // act
-        productInfoViewModel.updateProduct(product: nil)
+        sut?.updateProduct(product: nil)
         
         // assert
-        XCTAssertEqual(productInfoViewModel.weight, "Н/Д")
+        XCTAssertEqual(sut?.weight, "Н/Д")
     }
     
     func testThatPropertyCookingTimeReturnCorrectTimeOfCooking() {
@@ -86,10 +73,10 @@ class ProductInfoViewModelTests: XCTestCase {
         let product = ProductFake()
         
         // act
-        let productInfoViewModel = ProductInfoViewModel(product: product)
+        sut?.updateProduct(product: product)
         
         // assert
-        XCTAssertEqual(productInfoViewModel.cookingTime, "20 мин.")
+        XCTAssertEqual(sut?.cookingTime, "20 мин.")
     }
     
     func testThatProductInfoStackIsHiddenWhenProductIsNil() {
@@ -97,11 +84,11 @@ class ProductInfoViewModelTests: XCTestCase {
         let product: ProductProtocol? = nil
         
         // act
-        let productInfoViewModel = ProductInfoViewModel(product: product)
+        sut?.updateProduct(product: product)
         
         // assert
-        XCTAssertNil(productInfoViewModel.product)
-        XCTAssertTrue(productInfoViewModel.isHiddenProductStackView)
+        XCTAssertNil(sut?.product)
+        XCTAssertTrue(sut?.isHiddenProductStackView ?? false)
     }
     
     func testThatInitializationOfProductIsWorking() {
@@ -120,24 +107,24 @@ class ProductInfoViewModelTests: XCTestCase {
         let product = ProductFake()
         
         // act
-        let productInfoViewModel = ProductInfoViewModel(product: product)
-        let timerViewModel = productInfoViewModel.getTimerViewModel()
+        sut?.updateProduct(product: product)
+        let timerViewModel = sut?.getTimerViewModel()
         
         // assert
-        XCTAssertEqual(timerViewModel.minutes, 20)
+        XCTAssertEqual(timerViewModel?.minutes, 20)
     }
     
     func testThatCurrentStateIsChangeToFirstStateIfProductIsNil() {
         // arrange
         let product = ProductFake()
-        let productInfoViewModel = ProductInfoViewModel(product: product)
+        sut?.updateProduct(product: product)
         var currentState: String?
         let firstCompletion  = { currentState = "firstState" }
-        productInfoViewModel.needUpdateViewForFirstStep = firstCompletion
+        sut?.needUpdateViewForFirstStep = firstCompletion
         
         // act
-        productInfoViewModel.updateProduct(product: nil)
-        productInfoViewModel.checkCurrentStateAndUpdateView()
+        sut?.updateProduct(product: nil)
+        sut?.checkCurrentStateAndUpdateView()
         
         // assert
         XCTAssertNotNil(currentState)
@@ -147,14 +134,13 @@ class ProductInfoViewModelTests: XCTestCase {
     func testThatCurrentStateIsChangeToSecondState() {
         // arrange
         let product = ProductFake()
-        let productInfoViewModel = ProductInfoViewModel()
         var currentState: String?
         let secondCompletion = { currentState = "secondState" }
-        productInfoViewModel.needUpdateViewForSecondStep = secondCompletion
+        sut?.needUpdateViewForSecondStep = secondCompletion
         
         // act
-        productInfoViewModel.updateProduct(product: product)
-        productInfoViewModel.checkCurrentStateAndUpdateView()
+        sut?.updateProduct(product: product)
+        sut?.checkCurrentStateAndUpdateView()
         
         // assert
         XCTAssertNotNil(currentState)
@@ -164,14 +150,14 @@ class ProductInfoViewModelTests: XCTestCase {
     func testThatCurrentStateIsChangeToThirdState() {
         // arrange
         let product = ProductFake()
-        let productInfoViewModel = ProductInfoViewModel(product: product)
+        sut?.updateProduct(product: product)
         var currentState: String?
         let thirdCompletion = { currentState = "thirdState" }
-        productInfoViewModel.needUpdateViewForThirdStep = thirdCompletion
+        sut?.needUpdateViewForThirdStep = thirdCompletion
         
         // act
-        productInfoViewModel.buttonStartCookTapped = true
-        productInfoViewModel.checkCurrentStateAndUpdateView()
+        sut?.buttonStartCookTapped = true
+        sut?.checkCurrentStateAndUpdateView()
         
         // assert
         XCTAssertNotNil(currentState)
@@ -181,19 +167,19 @@ class ProductInfoViewModelTests: XCTestCase {
     func testThatFuncCellViewModelReturnCorrectInstructionCellViewModel() {
         // arrange
         let product = ProductFake()
-        let productInfoViewModel = ProductInfoViewModel(product: product)
+        sut?.updateProduct(product: product)
         let indexPath = IndexPath(row: 2, section: 0)
         let instructionText = Inscriptions.instructionOfCookingThirdStep
         
         // act
-        let cellViewModel = productInfoViewModel.cellViewModel(at: indexPath)
+        let cellViewModel = sut?.cellViewModel(at: indexPath)
         
         // assert
         XCTAssertEqual(cellViewModel?.instrImage, "instr2")
         XCTAssertEqual(cellViewModel?.numberOfCard, "3")
         XCTAssertEqual(cellViewModel?.getInstrLabel(), instructionText)
         XCTAssertNotNil(cellViewModel?.isShowNextLabel)
-        XCTAssertEqual(cellViewModel?.isShowNextLabel, false)
+        XCTAssertFalse(cellViewModel?.isShowNextLabel ?? true)
     }
     
 }
